Scaffolded instead

+++ db/migrate/20141109012351_create_snacks.rb
@@ -2,7 +2,7 @@ class CreateSnacks < ActiveRecord::Migration
   def change
     create_table :snacks do |t|
       t.string :name
-      t.float :calories
+      t.integer :calories
 
       t.timestamps
     end
